Fixed: Don't try to clean Update Folder if it doesn't exist
(cherry picked from commit 24f96fa2a4c1fd92f6433ec94edf6e55b7c83d89)

+++ src/NzbDrone.Core/Housekeeping/Housekeepers/CleanupTemporaryUpdateFiles.cs
@@ -1,4 +1,4 @@
-﻿using NzbDrone.Common.Disk;
+using NzbDrone.Common.Disk;
 using NzbDrone.Common.EnvironmentInfo;
 using NzbDrone.Common.Extensions;
 
@@ -19,7 +19,10 @@ namespace NzbDrone.Core.Housekeeping.Housekeepers
         {
             var updateSandboxFolder = _appFolderInfo.GetUpdateSandboxFolder();
 
-            _diskProvider.DeleteFolder(updateSandboxFolder, true);
+            if (_diskProvider.FolderExists(updateSandboxFolder))
+            {
+                _diskProvider.DeleteFolder(updateSandboxFolder, true);
+            }
         }
     }
 }
